Tak/hnt 1165 show everyone role for token gated towns (#1962)
When the user creates a token-gated space, the space contract creates 2
roles behind the scene - the Member role, and an Everyone role. The
Everyone role has 0 permissions. Previously, any role with 0 permissions
are hidden from view.
With this change, all roles in the space will be shown. We want to show
the Everyone role because it allows the admin to create a channel
which is readonly for Everyone.