MaxMember quota with allowlist for space joins (#1983)
Closes HNT-1098



In this PR:
- query to read current room member count in Dendrite
- gating in Dendrite `performJoin` handler to compare current count with
quota
- updated client integration to test local quota (5 max) config

In a subsequent PR:
- update SpaceJoin.tsx ( cc @salzbrenner ) component and add a new
MatrixError in the process to handle max quota related error on Invite
flow. Right now only missing token errors are handled in the invite
modal so until this is done users will see that error when max quota is
reached.

To allow Pioneer Town (and any other spaceId we want to avoid the max
cap), we can append the full spaceId to dendrite global config
`excluded_towns` array.