Sample App: enable casablanca sign in and create space (#1980)
Added button to the login screen to log in with casablanca. 
Added a dropdown on the create space page to pick a protocol.

Added a route to the logins page so that you can log into casablanca
after you’re logged into matrix, fixed up the page so it renders nice
when you’re logged in.