app: fix the defaultValues for channel settings form (#2050)
Sometimes the channel settings form displayed incorrect, stale data for
the checked roles, that maybe (?) was sometimes different than the
actual data submitted. The main cause of this was the `defaultValues`
passed to the form:
1. initial `defaultValues` had an incorrect check for which roles apply
to channel (and should be checked)
2. react-hook-form caches the values passed to `defaultValues` on first
render, and doesn't updated the form if the reference later changes. You
have to [reset](https://react-hook-form.com/api/useform/reset/) the
fields when the default values change. In this case, the fetched role
details might come in on subsequent renders.

With those changes, mostly everything worked as expected. The rest of
this PR is mostly separating code into files, and adding tests for the
form.

Additionally, this PR includes invalidating the react-query on the
create channel form as well - needed in the case of creating a channel
then immediately editing it.