app: allow user to upload image on create town form (#2053)
This PR adds image uploads to the create town form. The user picks their
image (optional) while in the form, and once mint is successful, it
uploads the image in the background. If there's an error with the
upload, it removes the image and notifies the user that the upload
failed via toast.

- we had a couple variants of uploading components. Consolidated them
into `LargeUploadImageTemplate` and `SmallUploadImageTemplate`, so all
upload components now use the same hooks/stores
- created a `useImageOnChangeEvent` hook for upload behavior
- Added a `FailedUploadAfterSpaceCreation` toast

Recordings:
[image success](https://jam.dev/c/9ecad2f0-1638-4001-9891-0d7a000887d6)
[image failure](https://jam.dev/c/8dfd4905-e5b9-4768-9849-5a8020406cc1)