Update the way our clients handle chainId changes (#2071)
This continues work that Evan has already done, he got us 95% of the way
there.

This change makes chainId a required paramter to the zion client. Any
changes to the chainId will require an app refresh.

I’m trying to create a system where we can be logged into casablanca and
matrix at the same time. In order to do that we need to make certian
assumptions about chains and server urls

1. chains are tied to server urls. If we’re on a localhost server, the
only chain that we should be looking at is foundry
2. we can only be on one chain at a time. it doesn’t make sense to be
logged into localhost casablanca and prod matrix.
3. logically from 1 and 2, the environment should consist of { chainId,
matrixUrl, casablancaUrl }, any time we switch environment, we should
reload the page and create a new zion client

By thinking about things this way I was able to remove a few hacks and
odd corners

In a follow up I’m re structuring the sample app to have an environment
dropdown that lets us switch between local, test, and prod.