Fix hnt-1358 user should be logged out if the app is not connected to any wallet accounts (#2143)
Issue: when the user disconnects all wallet accounts in MetaMask, the
app is still logged in. The login states are cached, and wagmi does not
tell the app that it is no longer connected to any wallets.

Fix: implement a custom InjectedConnector to receive all accountsChanged
callbacks. Then emit events to let our hooks handle the change.