Force userId to be required in Mentions (#2165)
I’m typing out the new payloads protobufs
https://github.com/HereNotThere/harmony/pull/2158 and it was questioned
why I would ever have an optional mention.

This seemed valid.

It wasn’t a required property before because there is some nuanced
parsing logic in the mention text node that didn’t take it into
consideration.

I gave this my best shot. The tests were able to catch issues… for
instance in the tests the match always has an @ symbol, but the lookup
doesn’t, so that was always failing.

But, I’m not sure how to test the serialization, or if it event matters
if serialization doesn’t result in a userId

Mentions currently work on staging, so some code path here was valid.

I tested that mentioning a user works on localhost between firefox and
chrome.