Reenable switching chains (#2164)
When I switch chains locally to a chain that I’m currently connected to,
I get this in the data paramater in the callback in useWeb3

```
{
    "chain": {
        "id": 5,
        "unsupported": false
    }
}
```

there’s no `data.account`,
Before this pr:
i just get stuck with no active wallet even though I’m connected
After this pr:
active wallet is updated to the correct wallet. 

Why this is necessary:
We added code that handled the disconnect action from a wallet, but it
was overly broad, and created a bug when the user switched chains.