Add a watch to generate go files from protobuf changes (#2168)
I intentionally did not add this to the start local dev script, since
these files are checked in and should be up to date on all branches

It should be useful for anyone working with the protos however. I will
run it while I’m tinkering with things.

Tested by starting the script, changing protos, and observing the
generated go files.