Refactor signer out of zionclient (#2237)
Everywhere we used a signer it’s now passed in as an argument. A
“correct” and up to date signer now lives in the web3Context.

Problem: if the signer changed during the lifetime of our client, the
wagmi signer object would update, but we wouldn’t refresh the zion
client or any of the space daps references to the previous signer

Solution: only require a signer when we need to write to a contract,
always pass it in

Additional solution
- signer is instantiated with the "appChainId" (the chainId passed into
the context via the environment)
- signer is held along side Chain in the web3 context
- I tried to mock the signer in the tests, but failed, so we still have
the option to pass in a signer.

Please Note:
As Tak pointed out in the PR review, this change does not guarantee that
the signer address matches our logged in wallet address. Additional work
is required to address this issue.