Jt/hnt 1434 device key upload (#2211)
Closes HNT-1434, HNT-1433. 
Update: 05/17: Merged #2245 to this branch, which along with this PR
Closes HNT-1434, HNT-1433.

This PR implements support for **device key upload** building off of
proto changes in #2190 that introduced new user stream
**UserDeviceKeyStream** that will store device_keys associated with
`device_ids` as well as `fallback_keys`, which will be used in lieu of
`one_time_keys` in Casablanca to establish an initial p2p Olm session
with a new peer.

As part of the work in this PR, I started porting the `Crypto` module
from matrix to Casablanca stripping out functionality to only what is
needed for this PR (for now). The intention is that subsequent follow on
PRs will then add necessary functionality + types around encryption -
incrementally adding to the Crypto module class. For some context, in
the current app, the Crypto module is `created` and `initialized` within
`startMatrixClient...` callback once a user a) connects their wallet to
the app b) registers with matrix ( to obtain a userId, deviceId...).
That library along with `MatrixDecryptionExtensions.ts` are the core
modules that support e2e encryption in the current app and will be
needed in Casablanca given our decision to port olm/megolm encryption
libraries that these modules encapsulate.

The Crypto class (matrix impl) loosely does the following:
1. loads the olm library
2. sets verification methods
3. sets crypto callbacks
4. instantiates crypto local storage to store key material
5. initializes event listeners for device key, encryption lifecycle
events
6. initializes olmDevice
7. registers deviceKeys for deviceId using olmDevice
8. uploads deviceKeys (ed25519, curve25519) to homeserver
9. starts tracking device List for joined rooms members' device list
changes

In this PR, to support uploading device keys to a new user_id prefixed
stream I've:
 - [x] created new Crypto class as TypedEmitter
- [x] signObject stub method ( needs olmDevice to be implemented in
class )
- [x] uploadDeviceKeys method that calls into client to upload signed
device keys to node
- [x] client implementation that creates or loads user device key stream
on startup
- [x] associated types and helpers for Crypto class
- [x] node checks on payload prior to adding a new Event ( note: needs
actual signed deviceKeys to `authenticate` )
- [x] primordial deviceId injection on Casablanca client startup (note:
this is needed for the Crypto class and not the finasl deviceId we're
going to eventually employ, which will be based on our encryption
related specs on
[deviceKeys](https://www.notion.so/herenottherelabs/RFCs-ba5254ccbab140afa1009a68b26cf7ab?p=cf08a116a724443f9a17b9558190db77&pm=s)
and implemented as follow-ons to this).
- [x] passing client test that uploads device key on client startup

Follow-on work after this PR is merged related to deviceKeys in
Casablanca:
1.
https://linear.app/hnt-labs/issue/HNT-1455/authenticate-user-is-writing-to-their-userdevicekey-stream-on-node
2.
https://linear.app/hnt-labs/issue/HNT-1450/implement-cryptostore-for-devicekeys
3.
https://linear.app/hnt-labs/issue/HNT-1451/implement-devicelist-trackers-for-joinedrooms
4.
https://linear.app/hnt-labs/issue/HNT-1452/port-olmmegolm-to-casablanca-crypto-module
5.
https://linear.app/hnt-labs/issue/HNT-1453/ensure-key-sharing-in-matrixextensiondelegate
6. https://linear.app/hnt-labs/issue/HNT-1454/devicekey-lifecycle