OlmMegolm delegate in Casablanca with tests (#2284)
Closes HNT-1417

This PR finishes migrating olm wasm
[library](https://gitlab.matrix.org/matrix-org/olm/-/tree/master/) to
our new `mecholm` (thanks @texuf ) package in casablanca. The only
object from the original library that is not exposed by the delegate
class as of this PR is SAS (short authentication string) which is an
opt-in [add
on](https://matrix.org/docs/guides/implementing-more-advanced-e-2-ee-features-such-as-cross-signing)
for emoji verification. We can add this in the future.

In this Pr:
- [x] delegate class that wraps Olm library, exposing only those objects
that are needed for porting the Crypto module (basically everything in
the Olm namespace minus SAS)
- [x] ported unit tests from Olm library that run against delegate class
- [x] more initialization code in our Crypto module that inits the
OlmDelegate, which needs to be done once for each client instance.

Subsequent to this PR (not in a specific order), we should finish
initializing deviceKeys in Crypto module, implement device list
tracking, starting hooking up event listeners to Casablanca client and
establish our cryptoStores in the Crypto module to persist session /
device keys.

---------

Co-authored-by: Kerem Kazan <kerem.kazan@gmail.com>