hnt-1382 still seeing channels despite not having tokens to join (#2318)
Issue: the state management on both the dendrite server and the
client are not in sync with the onchain states.
Fix on the dendrite-side is to check the blockchain entitlement state on
incremental sync. If user is not allowed to "Read", then do a membership "leave"
operation.
Fix on the client-side is to re-sync the space hierarchy when the
transaction is completed. Added a lot of console.log to debug the states
when role settings are changed.