Refactor toDevice proto further typing for key requests/response (#2315)
This PR adds increased typing for `ToDevice` messages that arise on a
user's UserStream to facilitate key request / response exchanges. This
will help the client detect key requests from responses **prior** to
decrypting the event payload and also enforce stricter typing prior to
sending a toDevice message from the client.

Prior to this, we would use the single op field `TDO_TO_DEVICE_OP` or
all toDevice messages, which didn't add any informational value to the
payload since the message type was already labelled as "ToDevice". We
also we're loosely following a convention (see tests) adding a `type`
field to the payload value to further encode the type. With this PR, we
don't need to follow that convention anymore and can rely on OP fields.