hnt-1527 users can unexpectedly be removed from channels (#2333)
Fix regression that cause by prior PR that sends a "leave" request if the admin changed the gating rules.

Fixes on dendrite:
* RoomInfo has a bug that does not correctly return the creator of a
channel
* The previous fix triggers a "leave room" event if the user is no
longer entitled to stay in the channel. However, I had to revert this
change. Because it caused regressions in the e2e tests. The tests were
racing ahead before the blockchain states are up-to-date. Need to figure
out how to fix this at a later date.

Fixes on client:
* not 100% reproducible, when minting new towns, sometimes, the
"general" channel is lost. Also noticed it when printing out
spaceHierarchy states in console.log. Possibly due to promises returning
out-of-order.