hnt-1535 #general channel does not show in UI after town is minted (#2351)
Fixed as follows:
* in `CreateSpaceFormStep3`, the `onAnimationComplete` navigated away
before the transaction is completed successfully. Moved the navigate
logic to the `CreateSpaceForm` in a `onTransactionSuccessful` callback.
* in `useCreateSpaceTransaction`, force a syncSpace to refresh the
SpaceHierarchy when the transaction is completed successfully. This
refreshes the spaceHierarchy at the earliest time.
* the `useSpaceHierarchies` uses a state `inFlightCurrent` to track
which space is being sync'ed. This is a `Promise<void>` object. It was
difficult to print state and reason about correctness. Re-wrote it to
use a `currentSpaceId`. This tracks the currentSpaceId being sync'ed and
is set to null when the sync is done.