app: remove useAuth().connectLoading (#2415)
While starting to look at
[HNT-1377](https://linear.app/hnt-labs/issue/HNT-1377/improve-ux-when-user-has-disconnected-the-wallet-from-metamask-or-if)
I noticed an issue that I thought was resolved - the app hangs on the
loading screen occasionally. This time it was happening for me locally
and easier to investigate.

In the end, I think this may be a bug within Wagmi. I scoured their
issues and source code - tried what was suggested
[here](https://github.com/wagmi-dev/wagmi/discussions/402) just to make
sure. It's very finicky when this bug appears, and only does so when I
first open the site, sometimes.

Anyway, I think other areas of the app have progressed (see Suspense
fallback in Main.tsx), to where the loading screen is shown
appropriately, and we don't actually need the check I've removed in
App.tsx. And that was the only place where `connectLoading` was used, so
I removed it. Also checked the lib, it doesn't have any dependency on a
'reconnecting' status.