tests no longer pipe stdout & stderr (#2425)
This PR is an attempt to reduce logger noise while running e2e tests.
Otherwise each run generates about 200k lines of logs - which can't be
viewed from the browser. This way, we can at least see which tests are
failing, and rerun those specific ones on local without the --silent
option if need be.