app: only fetch contract channels when opening the browse modal (#2460)
We don't need to check the contract channels to show the "Browse
Channels" option in the space bar - don't know what I was thinking when
I put that in. I think b/c of case for when town was created but had no
channels. Now we have a #general channel created on town mint, so not an
issue anymore. Regardless, my implementation was not ideal.
Only querying channels now when user actually opens the list of channels
modal.