lib: require walletAddress for client.syncSpace and add useLoggedInWalletAddress (#2491)
Problem: 
On first load, the app seems to hang and never load, but it only happens
sometimes. See
[here](https://linear.app/hnt-labs/issue/HNT-1648/entitlement-check-filtering-channels).
While investigating, I realized we also had an oversight in that you
could load the app with the wallet you did not log in with, and
reproduce this behavior.

Solution:
The on-first load issue - I'm pretty certain this is because in rare
instances there was a race condition where the sync + entitlement check
ran before Wagmi had the connected account address (or Wagmi was just
being buggy), so the wallet is undefined, the entitlement check never
runs and the channels are all filtered out.

The different wallet issue, we need to lock the channel entitlement
check to the logged in wallet.

Grabbing the logged in wallet from credentials store and passing it in
should fix both issues. I created the `useLoggedInWalletAddress` because
the other hooks that have this info use `useZionContext` and this is
being used outside of that context.

I'll follow up this PR with an investigation into react query usage here
- though I don't believe it's caused these issues, if `syncMatrixSpace`
isn't being called that often, and refetching the entitlement checks
doesn't degrade UX, we could remove it.