add getting of account address from user Id (#2510)
I'm not sure if I like this approach so happy to discuss. The reason for
this new code is the following - there are places where we need account
address only (for example, in
[use-current-wallet-equals-signedin-accounts.ts](https://github.com/HereNotThere/harmony/blob/30c61f8d96621ccdcd8cfe16b0a258e41c735372/clients/web/lib/src/hooks/use-current-wallet-equals-signedin-account.ts#LL15C41-L15C41))
and to generate full user id in case of River we would need to use
createUserIdFromEthereumAddress which required chainId also, though
userId is sufficient for account address in River, so that is the reason
for this new code.

So, here is a tradeoff - additional function in user-identifier.ts vs
additional imports in use-current-wallet-equals-signedin-accounts.ts,
which, probably, should not be there.