lib: remove messageScrollbackHooks.test, update mesageHistoryHooks (#2524)
Originally, this issue was for flaky `messageScrollbackHooks.test`, but
is essentially the same as `messageHistoryHooks.test`. Am I missing
something?

Assuming not, since `messageScrollbackHooks.test` has been quite flaky,
I deleted it.

However, I found a flaky run for `messageHistoryHooks.test`
[here](https://github.com/HereNotThere/harmony/actions/runs/5299906766/jobs/9593155141).
So I debugged this test. With the current changes, I ran it 300+ times
w/o failures. The solutions I've implemented point to an issue with
Dendrite (_possibly_ + anvil?), but don't address the dendrite issues. I
didn't try to debug that.

The solution was:
1. Increase the 401 retries in `useWithCatch` in tests to 12. This is
not ideal, but there were repeated test failures with 401s, even with a
number like 6. So if we're adding test specific logic, why not a high
enough number to get them to pass.
2. Adding longer timeouts to the test. Makes sense if the retries are
taking a long time.

Side note:
Both of these `Hooks` tests have similar coverage as
`messageScrollback.test` (this is a different test than
`messageScrollbackHooks.test`. The difference is, the `Hooks` test also
ensures that the React hook (`useChannelTimeline`) is working correctly
with `scrollback()`. So perhaps this is a good case for mocking the
scrollback behavior and simply testing the timeline hook - but I'm not
certain how much mocking that would entail.