app: update intro + fancy button (#2601)
**Background:** last couple of months the we've shuffled around code
quite a lot in the login area, wallet issues, mobile version etc. The
flow has changed a bit too and the ability to "Disconnect" from a wallet
before proceeding to login added. All for the better but the login isn't
as "smooth" anymore, buttons are popping and moving while changing
state. This could be improved more but I've been trying to be careful
modifying the least of code since the area is sensitive and this feels
good enough for this iteration.

- Separate connected state and disconnected state
- Move out "Disconnect Wallet" from Connect Button and make it more
subtly, avoiding having two similarly looking buttons.
- Misc clean up / Tweak transitions + add Fancy button (notes below)
- Add dark / light theme to rainbow wallet


https://github.com/HereNotThere/harmony/assets/557990/d009bdb2-12b1-4c69-9cdc-365830f4b6f0


Also I'm adding a `FancyButton` - Instead of trying to hard making a
one-size-fits all Button, I'm thinking of separating the functional
swiss-knife version and the animated one. This way there's more room for
play without having to worry about breaking things or adding overhead.
The Fancy button is animating between states, rippling when tapping and
may even get a little hover glow at some point, however it's convoluted,
has tons of layers and listeners so I would only use it for exclusive
places such as login, save form, etc. On the other hand, the usual
`Button` is css only, animating with `:hover` has a palette of tones and
sizes and magically adapts to its surrounding.


https://github.com/HereNotThere/harmony/assets/557990/c14dccde-56b3-4096-bee6-124a4a8bb4ab

Playground Link:
https://test-harmony-web-pr-2601.onrender.com/playground/buttons