lib: in SyncSpace leave non-existent channels and shorten the cache for entitled channels (#2599)
You can create a channel and never finish the transaction, and close
your browser and we keep trying to sync and check entitlement on it on
subsequent visits. You are also effected if someone else follows the
same pattern in a space you belong to. (This probably happens for spaces
too)

This PR:
- checks that channel exists on chain
- leaves the channel if it does not AND you are a member (the creator)
- if not a member (different user created it) you will still try to
check entitlement until that user has left it and it no longer appears
in matrix hierarchy
- reduces the time for caching entitled channels and cleans up code for
related removeQueries