lib: fix tests that use SpaceContextProvider + useRoles/useSpacesForContract (#2578)
- Changes Jest's `maxWorkers` from 5 to 2. (See reasoning below).
- Fixes a bunch of transaction hook tests. Basically, these tests all
use `<SpaceContextProvider />` and or `<ChannelContextProvider />`, with
a space/channel that is potentially still transacting on blockchain, and
not ready for the rest of the hooks and/or assertions, which leads to a
lot of weird flakiness.

After implementing changes from second bullet point, tests pass reliably
locally, but almost always some of the newly unskipped tests fail in CI
because of the createSpace transaction. I'm not clear why more processes
cause the transaction to fail (contract code, test code, something with
anvil??), but setting workers to 2 has resulted in 5+ rounds of passed
tests in CI. It's not the ideal fix, but I propose this change for now,
so that we can continue working on lib code with more confidence,
instead of keeping all these tests skipped.