emit RiverEvent for new channel msg and fix tests (#2608)
This PR starts to alter client event emitters to emit RiverEvents as
early as possible ( specifically `channelNewMessage` event), prior to
adding the event to the client stream. As a consequence, tests needed to
be fixed. Using RiverEvents (versus ParsedEvents), which store
ciphertext and clear text event contents, will allow for the client ( in
next PR) to attempt decryption of events as they come off the wire.

Subsequent PRs:
1.  emit RiverEvent for toDevice messages similar to channel messages
2. attempt decryption of RiverEvents in client prior to adding event to
stream. Emit Decrypted event with outcome ( success or error )
3. Update CasablancaTimeline to listen for Channel messages versus
streamUpdated
4. Initialize CasablancaDecryptionExtension to listen for KeyRequests
and Decrypted events to "Start looking for keys" on errors.