Add node version check script to start local dev (#2593)
Checks node version, prompts you to update it if it's out of date I
didn't install it for you because you need to restart VSCode to get it
to take effect, and I don't know how to do that. According to stack
overflow you have to killall Electron, which seems a bit nuclear.