lib: make useEthersProvider actually use Alchemy (#2675)
in retrospect this was incredibly obvious 😑 
When I upgraded wagmi, i made it so that lib/spaceDapp were just calling
a public rpc, and we started hitting rate limits. Extra confusing b/c
app was still calling alchemy. I've mostly been developing against
foundry, so was not running into issues. @erikolsson and @ttillberg
sorry for not connecting the dots when you mentioned it before!