Fix the RegisterAndJoin component in TestComponents, update logs (#2712)
we were calling it with a new array every time so it was joining rooms
in a loop, oops. Casablanca was complaining that it was trying to
re-join. Never noticed because it would gracefully ignore the errors.

Update logs that aren’t matrix specific to not say things about matrix.