lib: add signMessageAbortController to use-credentials-store (#2727)
This PR adds an abort signal to the signature request in the login flow,
which will be useful for mobile sign in. The app will do something like
`signMessageAbortController.abort()` when appropriate.

On desktop, when you login and get a sign request, you can close the
wallet, or "Reject" it, and the error bubbles up as expected. On mobile,
with WalletConnect, if you close their "Open wallet" dialog, or if you
reject the sign from within the wallet, there's no error, our
signMessage() just hangs and the user is left with in a broken UI state
that requires reloading to get the login button working again.