app: handle mobile metamask sign ons (#2730)
Adds a unique flow for users signing in on mobile with Metamask. They
get an additional modal after hitting "login" which has text telling
them to make sure they're on the right network. If they're on the wrong
network and try to sign the message, Metamask won't do anything, so we
show them a reset message after 10 seconds. More details and video in
this
[thread](https://hntlabs.slack.com/archives/C054RMW4YAV/p1689807489001589?thread_ts=1689619836.480849&cid=C054RMW4YAV).

This PR includes a new toast notification, that I ultimately ended up
not using in favor of a simple text message, but I'm including it in
case we want to keep it.