Tak/hnt 1756 pnw fix mention logic to notify user 1 time only (#2733)
Original issue: @mentioned user receives 2 notifications. This is
because Dendrite has no visibility into the encrypted message content.
It sends a notification to all members of the channel. However, for
@mention, we wanted to show a targeted "so-and-so mentioned you
message". Hence, the client sends a direct POST request to the Push
Notification Worker (PNW) so that it can craft this message. Doing this
causes the 2 notifications problem.

Re-designed the logic. Instead of sending a second push notification
to the client, the PNW now records the @mention request as a "hint" in
its DB. When it receives a notification request from Dendrite, the PNW
checks its db for this hint record. If an @mention hint is present, it
changes the notificationType to NotificaitonType.Mention. That way, when
the client receives the notification, it is either
NotificationType.NewMessage, or NotificationType.Mention. The client
only displays 1 message.