Tak/hnt 1797 reply_to notification and future notification types (#2740)
Introduce the concept of NotificationTag. 'mention', 'reply_to' are
different kinds of tags. The client sends these tags to the Push
Notification Worker based on the user's actions. When a notification is
triggered, the PNW uses these tags as hint to figure out the right
NotificationType - NewMessage, Mention, or ReplyTo.

When the Service Worker receives and handles the push event, it can
switch on the NotificationType to show the right notification message to
the user.