lib: add initialSyncSortPredicate and timeBetweenSyncingSpaces to ZionContext (#2753)
Two additions:
1. From a UX standpoint, it would be nice to sync the space that you
initially load first, especially on mobile. The app has logic that on
load, it will route you back to the last space you were in when you
left. I'm in a bunch of spaces and it takes a while for ones down the
queue to sync, even though thats the one I want to see first. Hence the
`initialSyncSortPredicate`. I was going to add this to useCasablanca
spaces as well, but in light of the Aug 7 milestone and since I'm not
too familiar with it, I'll hold off for now.
2. `timeBetweenSyncingSpaces` is probably not going to solve what some
users experience when loading the app on mobile - the main thread
freezing up - but I figured we could try and find out.