app: more aggressive push subscription registration (#2782)
We seem to be missing subscriptions in the worker for some reason. Let's
make the registration a bit more aggressive.

**Changes**
- By moving the logic into a separate component
(`<RegisterPushSubscription>`) and placing it closer to the root, we can
get rid of the `useRef` dance we used previously.
- we no longer check if we have registered the subscription before or
not. we always assume that the worker hasn't stored our latest push
subscription.
- `usePushNotifications` is now responsible only for showing the banner,
and `usePushSubscription` handles the interaction with the worker.

The app will now notify the worker exactly once on every app
load/notification permission granted/user id change, regardless of
whether it has notified the worker before or not.