Reduce GetStream noise (#2795)
Reduce noise on River node when running tests. GetStream would print an
error for streams that do not exist. It now prints a debug log. Also, in
tests, we used to invariably attempt to loadUser stream even when
registering a new user and catch failure to then create new User
streams. This was due to the fact we don't return a cookie from River
node that indicates whether the user exists or not. Let's avoid an
unnecessary GetStream call over rpc, when we know we are registering a
new client as we do in tests.