Blockheader Snapshots, data types and server side generation (#2831)
I don’t have - blockheader_next.go, which would create the next block
from the previous block, but that should be pretty easy
And I haven’t touched the client, that’s going to be harder than it
seems, and i’d like to wait until i know a little more about how they
will sync to the client. It feels like the response object is going to
have to change.