app: integrate notification settings api (#2866)
In this PR:
- removes the useMuteSettings zustand store and service worker logic
- hooks up the UI that was using the store to call the push worker api

@tak-hntlabs I kept all this in the app for now. We can decide where we
want to put it later. I'm really in favor of moving all the push worker
stuff out of lib and extending the scope of zionClient._eventHandlers to
trigger events as needed.

I've confirmed that all the UI behaves as it was previously. Confirmed
push notifications do not come through in following situations:
- user has muted channel
- user has muted space - no notifications from any channel
- user has left channel
- user has left space

---------

Co-authored-by: Tak Wai Wong <64229756+tak-hntlabs@users.noreply.github.com>