preliminary version of use-casablanca-store.ts (#2875)
It is preliminary implementation of use-casablanca-store.ts which is an
equivalent of Matrix related
[use-matrix-store.ts](https://github.com/HereNotThere/harmony/blob/2b60419199395615219ce1cba33042fa9117a812/clients/web/lib/src/store/use-matrix-store.ts#L4)

It is needed to complete mentions and will be used inside of
[useCasablancaRooms.ts](https://github.com/HereNotThere/harmony/blob/2b60419199395615219ce1cba33042fa9117a812/clients/web/lib/src/hooks/ZionContext/useCasablancaRooms.ts#L5)

There is still bunch of TODOs to proper fill necessary entities, though
now I'm concentrated on users joined spaces and channels. Logic of
filtering should be validated during review and testing.