HNT-1837: Tighten river event types from any (#2871)
HNT-1837 - This PR goes a long way to tightening RiverEvent types from
essentially `any` to `PlainMessages` of the concrete type expected by
RiverEvents. The way I think about the RiverEvent model is that it needs
to bi-directionally support encrypting input and decrypting output, all
with types for each input / output. It shouldn't support more types than
the expect input/output set. As of now, e2e encryption and therefore
RiverEvent covers ToDevice and Channel messages defined in
payloads.proto and protocol.proto. This PR tightens the
`IContent.content` defined in event.ts type with this in mind.

Most other changes in this PR are to rightsize tests that broke and
employ new typeguards to allow callers to retrieve typed content
depending on whether the RiverEvent is seen over to device or channel
transport and whether the event is presumed encrypted or decrypted into
a clear event.

As a follow on in the same vein as the ticket , I'll be split
sendToDevice function and continue removing type casts that still linger
in event.ts.