app: fix clearing unreads on loading from notification (#2881)
you could have the window open and it would read `isWindowFocused:
false` from storage, and the hook wouldn't fire because there wasn't an
event, so the condition in `<MessageTimeline />` wouldn't be met to mark
it as read. probably an oversight to save this value to storage, we
don't need it right? or am i missing something?