River Decryption extension (#2911)
Closes HNT-1453, HNT-1938. 

This PR implements RiverDecryptionExtension, which is River counterpart
to MatrixDecryptionExtension that @texuf wrote. This completes e2e
encryption feature on Channel messages for River client and protocol. As
a follow on, we should implement support for all Channel payload types
(found in `payloads.proto`) and wire the timeline to listen and update
state on `eventDecrypted` events (see
[HNT-1937](https://linear.app/hnt-labs/issue/HNT-1937/support-all-channelmessage-payload-types-riverevent)
).

In DRAFT, while I fix timing out sdk tests. The event emitter is slower
now that we're passing RiverEvents along with StreamEvents to stream
constructor ( in order to emit RiverEvent types from an emitted
RiverEvent in StreamStateView ).