app: add `home` route on touch (#2926)
This PR adds a `/t/:town_id/home` route, which is the state where the
user is in a town, but not yet inside a channel. This state has no
counterpart on desktop so I believe we will need to introduce a
touch-only route here.

Another alternative would be to simply make this route `/t/:town_id`,
but it kind of collides with the functionality in `SpaceHome`. Any
thoughts?

<img width="414" alt="image"
src="https://github.com/HereNotThere/harmony/assets/352982/7cad5648-6dfb-485a-964c-2b35444b0c3a">