Fix encryption check when processing replaced messages in timeline diff (#2948)
og pr: https://github.com/HereNotThere/harmony/pull/2910
fixes ci failures

The original CI was green. This failure mode only happens when the
message initially fails to decrypt, but then gets decrypted via the
special MatrixDecryptionExtensions later on. Looks like in CI this is
close to 100% of the time, but PR's only trigger it 25ish%.