app: hide tabbar while keyboard is active (#2931)
Not sure if this is the best way to communicate between the components,
but here's one way to toggle the tab bar while the channel text editor
is active. LMK what you think!

**Open Questions**
- instead of using a context, do we want to explore using the [Visual
Viewport
API](https://developer.mozilla.org/en-US/docs/Web/API/VisualViewport#hiding_an_overlaid_box_on_zoom)
as suggested by @ttillberg ? what would be the trigger condition in this
case?
- do we always want to hide the tab bar while editing, or only while
editing in the channel? it might make sense to always hide it when the
keyboard is open for the most app-like feel.


https://github.com/HereNotThere/harmony/assets/352982/f5609079-082b-4329-a694-af67f32f0213