app: vlist rewrite for momentum scrolling  (#2522)
rewrite of vlist (message timeline) primarily to work with touch /
momentum scrolling.

if you are reviewing beware `VList.tsx` (main file) is collapsed.

main changes:
- decouple "react state" and "scroll state" (imperative) - the latter
being updated on a frame basis will access dom through refs.
- readjusting scroll position via scrollBy when items are added, remove
or resized is necessary to keep the list stable. On desktop we can do
this in realtime via `scrollBy` however on mobile this technique kills
momentum scrolling. Instead we offset the list via `absolute`
positioning, adding up a dept that can be released and readjusted after
when the scrolling is over.
- tighter control over resizing via central resize observer

there are still imperfections but the imperative access to dom/scroll
makes it a lot more predictable and in sync, especially for operations
that needs to be handled the same frame, i.e. the resizing of an item
will not postpone reordering and rendering until the next frame
preventing flashes. An example of things this could allow is smooth
collapse/expand/delete/add animations.